[SPOT-301][FEATURE] 소셜 로그인 관련 DTO 패키지 정리
