+++ src/main/java/com/example/spot/auth/infrastructure/client/google/GoogleApiClient.java
@@ -2,7 +2,7 @@ package com.example.spot.auth.infrastructure.client.google;
 
 import static com.example.spot.auth.infrastructure.constants.AuthConstants.HEADER_AUTHORIZATION;
 
-import com.example.spot.auth.presentation.dto.GoogleUser;
+import com.example.spot.auth.presentation.dto.oauth.google.GoogleUser;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
