+++ src/main/java/com/example/spot/auth/infrastructure/client/kakao/KaKaoApiClient.java
@@ -3,7 +3,7 @@ package com.example.spot.auth.infrastructure.client.kakao;
 import static com.example.spot.auth.infrastructure.constants.AuthConstants.HEADER_AUTHORIZATION;
 import static com.example.spot.auth.infrastructure.constants.AuthConstants.HEADER_CONTENT_TYPE;
 
-import com.example.spot.auth.presentation.dto.KaKaoUser;
+import com.example.spot.auth.presentation.dto.oauth.kakao.KaKaoUser;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
